Switch to a custom base64 implementation
According to the official specification, string encoding is not using
standard url-safe Base64 encoding, but a custom format that works
on 6-bit aligned data, without padding.

While the existing code in this project worked on the published examples
and reference online encoder/decoder, it failed when applied on the
updated official examples, in particular the header values.