Add experimental US TN implementation
Unit tests will be added when the iabgpp.com official website adds
support for this section so we can generate sample payloads.

Note that at the time this commit has been made, the
`MspaCoveredTransaction` field is missing from the specification. It
seems to be a bug, and I submitted an issue:
https://github.com/InteractiveAdvertisingBureau/Global-Privacy-Platform/issues/135