Add support for enums in iab_gpp_derive
This allows us to support sections where new versions are incompatible
with older ones.
The idea is for example to make the `Core` an enum with variants V1 and
V2, etc.

Note that it also means that we need to update error reporting, as there
is officially no "expected version" if several of them can be
encountered in the wild.
As a result, a lot of tests need to be updated.