Fix TCF CA field types
Core fields VendorExpressConsent and VendorImpliedConsent are documented
to be of type OptimizedRange. However, The Java implementation uses the
same encoder/decoder as for TCF EU v2, which uses the OptimizedIntRange
type.
As consequence, some payloads generated by the iabgpp.com tool failed to
decode in this library.
This commit switches the TCF CA fields to use OptimizedIntRange.