Refactor bit stream handling
Rationale: we don't want to maintain an extra layer between the bitsteam
implementation and the GPP Core data types.
Ideally, all custom types should extend `FromBitStream`.
The end goal are:
- stream directly from Base64 encoded strings so we don't allocate
memory for the binary buffer
- improve error reporting with context

Removes `DataReader` and `FromDataReader` types to rely on
`FromBitStream`.

The `BitRead` trait is extended with methods to parse GPP Core types.